requirements.txt
Django~=2.0.6
Django~=2.2
django-ckeditor~=5.7